% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{AcousticStudy-class}
\alias{AcousticStudy-class}
\title{\code{AcousticStudy} Class}
\description{
An S4 class storing acoustic data from an entire AcousticStudy
}
\section{Slots}{

\describe{
\item{\code{id}}{a unique id for the study}

\item{\code{events}}{a list of \linkS4class{AcousticEvent} objects with
detections from the AcousticStudy}

\item{\code{files}}{a list of folders and files containing the AcousticStudy data}

\item{\code{gps}}{a data frame of gps coordinates for the entire AcousticStudy}

\item{\code{pps}}{the \linkS4class{PAMpalSettings} object used to create this object}

\item{\code{settings}}{a named list of various settings for detectors, localizers, etc.}

\item{\code{effort}}{something about effort lol}

\item{\code{models}}{a place to store any models run on your data}

\item{\code{ancillary}}{miscellaneous extra data}
}}

\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
