% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStudy.R
\name{checkStudy}
\alias{checkStudy}
\title{Check an AcousticStudy Object for Issues}
\usage{
checkStudy(x)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} object}
}
\value{
invisibly returns \code{x} after printing possible issues
}
\description{
Checks for any possible issues in an \linkS4class{AcousticStudy}
  object, printing statements to the console
}
\examples{

data(exStudy)

# currently only checks if any peak frequencies are 0, so well force this
exStudy[[1]][[1]]$peak <- 0
checkStudy(exStudy)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
