% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exStudy}
\alias{exStudy}
\title{Example AcousticStudy Object}
\format{
a \linkS4class{AcousticStudy} object containing two
  \linkS4class{AcousticEvent} objects
}
\usage{
data(exStudy)
}
\description{
An example AcousticStudy object created using the example
PAMpalSettings object provided with the package. Processed with mode='db'
}
\keyword{datasets}
