% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWarnings.R
\name{getWarnings}
\alias{getWarnings}
\title{Get Warning Messages}
\usage{
getWarnings(x)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} or \linkS4class{AcousticEvent}
object}
}
\value{
a list of warning messages, named by the function call that created
  the warning
}
\description{
Accessor to easily get all warning messages for \code{x}
}
\examples{

\dontrun{
data(exStudy)
# This will trigger a warning, then we can access it
exStudy <- filter(exStudy, species == "test")
getWarnings(exStudy)
}

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
