% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{is.AcousticEvent}
\alias{is.AcousticEvent}
\title{Check if an Object is an AcousticEvent}
\usage{
is.AcousticEvent(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Function to check if an object is an AcousticEvent
}
