% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{is.AcousticStudy}
\alias{is.AcousticStudy}
\title{Check if an Object is an AcousticStudy}
\usage{
is.AcousticStudy(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Function to check if an object is an AcousticStudy
}
