% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{is.PAMpalSettings}
\alias{is.PAMpalSettings}
\title{Check if an Object is a PAMpalSettings}
\usage{
is.PAMpalSettings(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Function to check if an object is a PAMpalSettings
}
