% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMpalSettings.R
\name{PAMpalSettings}
\alias{PAMpalSettings}
\title{Constructor for PAMpalSettings Object}
\usage{
PAMpalSettings(db = NULL, binaries = NULL, calibration = NULL)
}
\arguments{
\item{db}{the full path to a PamGuard database file}

\item{binaries}{a list with items "folder" containing the directory of the
PamGuard binary files, and "list" containing the full path to each individual
binary file.}

\item{calibration}{a named list of calibration functions to apply while
applying functions from the "functions" slot. Should named by the
PamGuard module, same as the "functions"}
}
\value{
A PAMpalSettings object
}
\description{
Create a PAMpalSettings object. Any values that are not supplied
  will be asked for interactively.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
