% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSettings.R
\name{removeSettings}
\alias{removeSettings}
\title{Remove Settings from a PAMpalSettings Object}
\usage{
removeSettings(pps)
}
\arguments{
\item{pps}{a \linkS4class{PAMpalSettings} object to remove settings from}
}
\value{
the same \linkS4class{PAMpalSettings} object as pps, with all
  settings removed from the "settings" slot
}
\description{
Remove all settings from the "settings" slot in a PAMpalSettings
  object.
}
\examples{

exPps <- new('PAMpalSettings')
exPps <- addSettings(exPps, system.file('extdata', 'Example.xml', package='PAMpal'))
removeSettings(exPps)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
