% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardCepstrumCalcs.R
\name{standardCepstrumCalcs}
\alias{standardCepstrumCalcs}
\title{Calculate a Set of Measurements from a Cepstrum Contour}
\usage{
standardCepstrumCalcs(data)
}
\arguments{
\item{data}{a list that must have \code{quefrency} the "quefrency" at each
cepstrum contour, \code{sr} the sample rate of the data,
and \code{time} the time in seconds at each bin}
}
\value{
A list with inter-click interval (ici), duration (seconds), and
  slope of the ici
}
\description{
Calculate a set of measurements from a cepstrum contour. This
  is currently used to measure the inter-click interval of the burst pulse
  type calls
}
\examples{

data(testCeps)
standardCepstrumCalcs(testCeps)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
