% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testCeps}
\alias{testCeps}
\title{A fake cepstrum contour}
\format{
A list with three items:
\describe{
  \item{quefrency}{a vector of the cepstrum contour bin numbers, not actually quefrency}
  \item{time}{a vector of the time values of the cepsturm contour in seconds}
  \item{sr}{the sample rate of the recording}
}
}
\usage{
data(testCeps)
}
\description{
A manually created fake cepstrum contour, mimicing what the output
  would be from the Pamguard module and fed into the cepstrum calcs
}
\keyword{datasets}
