% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testClick}
\alias{testClick}
\title{A two-channel recording of a delphinid click}
\format{
A list with two items:
\describe{
  \item{wave}{a matrix with two columns of 800 samples, each column
  is a separate recording channel}
  \item{sr}{the sample rate of the recording}
}
}
\source{
Southwest Fisheries Science Center / NMFS / NOAA
}
\usage{
testClick
}
\description{
An example delphinid click waveform. This is a two-channel recording
of some kind of delphinid click, recorded at 192kHz. There are
800 samples recorded on each channel.
}
\keyword{datasets}
