% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testGPL}
\alias{testGPL}
\title{A fake GPL detection}
\format{
A list with three items:
\describe{
  \item{freq}{a vector of the frequency contour values in hertz}
  \item{time}{a vector of the time values of the contour in seconds}
}
}
\usage{
data(testGPL)
}
\description{
A manually created fake GPL contour, mimicing what the output
  would be from the Pamguard module and fed into the GPL calcs
}
\keyword{datasets}
