% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testWhistle}
\alias{testWhistle}
\title{A fake whistle contour}
\format{
A list with two items:
\describe{
  \item{freq}{a vector of the frequency contour values in hertz}
  \item{time}{a vector of the time values of the contour in seconds}
}
}
\usage{
data(testWhistle)
}
\description{
A manually created fake whistle contour reanging from 1kHz to 3.1kHz
}
\keyword{datasets}
