% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnnotation.R
\name{addAnnotation}
\alias{addAnnotation}
\alias{getAnnotation}
\title{Add Annotation Data to an AcousticStudy Object}
\usage{
addAnnotation(x, annoList = NULL)

getAnnotation(x = NULL, notes = FALSE)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} object}

\item{annoList}{an annotation list created by \code{getAnnotation}.
Default \code{NULL} will create a blank list}

\item{notes}{logical flag whether to pull a version of the list that only contains notes,
should only be \code{TRUE} for internal use}
}
\value{
The same object as \code{x} with an \code{$annotation} item added
  the the \code{ancillary} slot of each event in \code{x}
}
\description{
Adds annotation data to an \linkS4class{AcousticStudy}
  object, usually in preparation for exporting to an external source.
  Most pieces of the annotation form will be filled in by the user when
  the function is called, but \code{lat}, \code{lon}, \code{time_start},
  \code{time_end}, \code{freq_low}, \code{freq_high}, \code{source_id},
  and \code{annotation_id} will be filled in automatically based on data
  in each \linkS4class{AcousticEvent}. Annotations are stored for each
  event in the \code{ancillary} slot.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
