% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFunction.R
\name{addFunction}
\alias{addFunction}
\title{Add a Function to a PAMpalSettings Object}
\usage{
addFunction(pps, fun, module = NULL, verbose = TRUE)
}
\arguments{
\item{pps}{a \linkS4class{PAMpalSettings} object to add a function to}

\item{fun}{function to add OR another \linkS4class{PAMpalSettings} object.
In this case all functions from the second object will be added to \code{pps}}

\item{module}{PamGuard module output this function should act on}

\item{verbose}{logical flag to show messages}
}
\value{
the same \linkS4class{PAMpalSettings} object as pps, with the function
  \code{fun} added to the "functions" slot
}
\description{
Adds a new function to the "function" slot in a PAMpalSettings
  object.
}
\examples{

# not recommended to create a pps like this, for example only
pps <- new('PAMpalSettings')
if(interactive()) pps <- addFunction(pps, standardClickCalcs)
pps <- addFunction(pps, roccaWhistleCalcs, module='WhistlesMoans')

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
