% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRecordings.R
\name{addRecordings}
\alias{addRecordings}
\title{Add Recordings to an AcousticStudy Object}
\usage{
addRecordings(x, folder = NULL, log = NULL, progress = TRUE)
}
\arguments{
\item{x}{a \linkS4class{AcousticStudy} object to add recordings to}

\item{folder}{a folder of recordings to add}

\item{log}{(optional) log files for SoundTrap recordings}

\item{progress}{logical flag to show progress bars}
}
\value{
the same object as \code{x} with recording information added
  to the \code{files} slots
}
\description{
Adds recordings to an AcousticStudy object. Each
   AcousticEvent within
}
\examples{


}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
