% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStudy.R
\name{checkStudy}
\alias{checkStudy}
\title{Check an AcousticStudy Object for Issues}
\usage{
checkStudy(x)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} object}
}
\value{
invisibly returns \code{x} after printing possible issues
}
\description{
Checks for any possible issues in an \linkS4class{AcousticStudy}
  object, printing statements to the console
}
\examples{

# setting up example data
exPps <- new('PAMpalSettings')
exPps <- addDatabase(exPps, system.file('extdata', 'Example.sqlite3', package='PAMpal'))
exPps <- addBinaries(exPps, system.file('extdata', 'Binaries', package='PAMpal'))
exClick <- function(data) {
    standardClickCalcs(data, calibration=NULL, filterfrom_khz = 0)
}
exPps <- addFunction(exPps, exClick, module = 'ClickDetector')
exPps <- addFunction(exPps, roccaWhistleCalcs, module='WhistlesMoans')
exPps <- addFunction(exPps, standardCepstrumCalcs, module = 'Cepstrum')
exData <- processPgDetections(exPps, mode='db')

# currently only checks if any peak frequencies are 0, so well force this
exData[[1]][[1]]$peak <- 0
checkStudy(exData)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
