% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetSet.R
\name{PAMpal.accessors}
\alias{PAMpal.accessors}
\alias{settings}
\alias{settings,AcousticEvent-method}
\alias{settings<-}
\alias{settings<-,AcousticEvent-method}
\alias{localizations}
\alias{localizations,AcousticEvent-method}
\alias{localizations<-}
\alias{localizations<-,AcousticEvent-method}
\alias{id}
\alias{id,AcousticEvent-method}
\alias{id<-}
\alias{id<-,AcousticEvent-method}
\alias{detectors}
\alias{detectors,AcousticEvent-method}
\alias{detectors<-}
\alias{detectors<-,AcousticEvent-method}
\alias{species}
\alias{species,AcousticEvent-method}
\alias{species,AcousticStudy-method}
\alias{species<-}
\alias{species<-,AcousticEvent-method}
\alias{files}
\alias{files,AcousticEvent-method}
\alias{files<-}
\alias{files<-,AcousticEvent-method}
\alias{ancillary}
\alias{ancillary,AcousticEvent-method}
\alias{ancillary<-}
\alias{ancillary<-,AcousticEvent-method}
\alias{[,AcousticEvent,ANY,ANY,ANY-method}
\alias{[<-,AcousticEvent,ANY,ANY,ANY-method}
\alias{$,AcousticEvent-method}
\alias{$<-,AcousticEvent-method}
\alias{[[,AcousticEvent,ANY,ANY-method}
\alias{[[<-,AcousticEvent,ANY,ANY,ANY-method}
\alias{id,AcousticStudy-method}
\alias{id<-,AcousticStudy-method}
\alias{files,AcousticStudy-method}
\alias{files<-,AcousticStudy-method}
\alias{gps}
\alias{gps,AcousticStudy-method}
\alias{gps<-}
\alias{gps<-,AcousticStudy-method}
\alias{detectors,AcousticStudy-method}
\alias{events}
\alias{events,AcousticStudy-method}
\alias{events<-}
\alias{events<-,AcousticStudy-method}
\alias{settings,AcousticStudy-method}
\alias{settings<-,AcousticStudy-method}
\alias{effort}
\alias{effort,AcousticStudy-method}
\alias{effort<-}
\alias{effort<-,AcousticStudy-method}
\alias{pps}
\alias{pps,AcousticStudy-method}
\alias{pps<-}
\alias{pps<-,AcousticStudy-method}
\alias{ancillary,AcousticStudy-method}
\alias{ancillary<-,AcousticStudy-method}
\alias{models}
\alias{models,AcousticStudy-method}
\alias{models<-}
\alias{models<-,AcousticStudy-method}
\alias{[,AcousticStudy,ANY,ANY,ANY-method}
\alias{[<-,AcousticStudy,ANY,ANY,ANY-method}
\alias{$,AcousticStudy-method}
\alias{$<-,AcousticStudy-method}
\alias{[[,AcousticStudy,ANY,ANY-method}
\alias{[[<-,AcousticStudy,ANY,ANY,ANY-method}
\title{\code{AcousticEvent} and \code{AcousticStudy} accessors}
\usage{
settings(x, ...)

\S4method{settings}{AcousticEvent}(x, ...)

settings(x) <- value

\S4method{settings}{AcousticEvent}(x) <- value

localizations(x, ...)

\S4method{localizations}{AcousticEvent}(x, ...)

localizations(x) <- value

\S4method{localizations}{AcousticEvent}(x) <- value

id(x, ...)

\S4method{id}{AcousticEvent}(x, ...)

id(x) <- value

\S4method{id}{AcousticEvent}(x) <- value

detectors(x, ...)

\S4method{detectors}{AcousticEvent}(x, ...)

detectors(x) <- value

\S4method{detectors}{AcousticEvent}(x) <- value

species(x, ...)

\S4method{species}{AcousticEvent}(x, ...)

\S4method{species}{AcousticStudy}(x, type = "id", ...)

species(x) <- value

\S4method{species}{AcousticEvent}(x) <- value

files(x, ...)

\S4method{files}{AcousticEvent}(x, ...)

files(x) <- value

\S4method{files}{AcousticEvent}(x) <- value

ancillary(x, ...)

\S4method{ancillary}{AcousticEvent}(x, ...)

ancillary(x) <- value

\S4method{ancillary}{AcousticEvent}(x) <- value

\S4method{[}{AcousticEvent,ANY,ANY,ANY}(x, i)

\S4method{[}{AcousticEvent,ANY,ANY,ANY}(x, i) <- value

\S4method{$}{AcousticEvent}(x, name)

\S4method{$}{AcousticEvent}(x, name) <- value

\S4method{[[}{AcousticEvent,ANY,ANY}(x, i)

\S4method{[[}{AcousticEvent,ANY,ANY,ANY}(x, i) <- value

\S4method{id}{AcousticStudy}(x, ...)

\S4method{id}{AcousticStudy}(x) <- value

\S4method{files}{AcousticStudy}(x, ...)

\S4method{files}{AcousticStudy}(x) <- value

gps(x, ...)

\S4method{gps}{AcousticStudy}(x, ...)

gps(x) <- value

\S4method{gps}{AcousticStudy}(x) <- value

\S4method{detectors}{AcousticStudy}(x, ...)

events(x, ...)

\S4method{events}{AcousticStudy}(x, ...)

events(x) <- value

\S4method{events}{AcousticStudy}(x) <- value

\S4method{settings}{AcousticStudy}(x, ...)

\S4method{settings}{AcousticStudy}(x) <- value

effort(x, ...)

\S4method{effort}{AcousticStudy}(x, ...)

effort(x) <- value

\S4method{effort}{AcousticStudy}(x) <- value

pps(x, ...)

\S4method{pps}{AcousticStudy}(x, ...)

pps(x) <- value

\S4method{pps}{AcousticStudy}(x) <- value

\S4method{ancillary}{AcousticStudy}(x, ...)

\S4method{ancillary}{AcousticStudy}(x) <- value

models(x, ...)

\S4method{models}{AcousticStudy}(x, ...)

models(x) <- value

\S4method{models}{AcousticStudy}(x) <- value

\S4method{[}{AcousticStudy,ANY,ANY,ANY}(x, i)

\S4method{[}{AcousticStudy,ANY,ANY,ANY}(x, i) <- value

\S4method{$}{AcousticStudy}(x, name)

\S4method{$}{AcousticStudy}(x, name) <- value

\S4method{[[}{AcousticStudy,ANY,ANY}(x, i)

\S4method{[[}{AcousticStudy,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{a \linkS4class{AcousticEvent} or \linkS4class{AcousticStudy} object}

\item{\dots}{other arguments to pass to methods}

\item{value}{value to assign with accessor}

\item{type}{species type to select}

\item{i}{index of the object to access}

\item{name}{name of the object to access}
}
\value{
\describe{
  \item{id}{a unique id or name for this object}
  \item{settings}{a named list of settings for each detector and localization or recorder}
  \item{detectors}{a list of detector data frames}
  \item{localizations}{list of localizations}
  \item{species}{list of species classifications}
  \item{files}{list of files used to create this object}
  \item{events}{a list of \linkS4class{AcousticEvent} objects}
  \item{gps}{a dataframe containing gps data}
  \item{pps}{the \linkS4class{PAMpalSettings} object used to create this}
  \item{effort}{something about effort?}
  \item{ancillary}{miscellaneous extra data}
  }
}
\description{
Accessors for slots in \linkS4class{AcousticEvent}
  and \linkS4class{AcousticStudy} objects
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
