% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGram.R
\name{plotGram}
\alias{plotGram}
\title{Plot Spectrogram or Cepstrogram}
\usage{
plotGram(
  x,
  start = 0,
  end = NULL,
  channel = 1,
  wl = 512,
  hop = 0.25,
  mode = c("spec", "ceps"),
  detections = c("whistle", "click", "cepstrum")
)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} object}

\item{start}{start time of the plot, either POSIXct or seconds from
the start of the recording}

\item{end}{end time of the plot, either POSIXct or seconds from the start
of the recording.}

\item{channel}{channel to plot}

\item{wl}{window length of FFT to use for creating plot}

\item{hop}{hop value of FFT to use for creating plot, either as a percentage
of \code{wl} or number of samples}

\item{mode}{one of \code{'spec'} or \code{'ceps'} to plot either spectrogram
or cepstrogram}

\item{detections}{vector containing any of \code{'whistle'}, \code{'click'},
and/or \code{'cepstrum'} indicating which detections to overlay on the plot}
}
\value{
nothing, just plots
}
\description{
Plots either a spectrogram or cepstrogram and also overlays
  whistle or cepstral contours from the binary files
}
\examples{

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
