% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIciReview.R
\name{runIciReview}
\alias{runIciReview}
\title{Run Echo Depth Review App}
\usage{
runIciReview(x, maxIci = 2.5)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} object that has been processed with
\link{calculateEchoDepth}}

\item{maxIci}{maximum ICI value (seconds) to display on plots}
}
\value{
the object as \code{x}, with updated \code{keepClick} column
}
\description{
Runs a Shiny app
}
\examples{
# example not run because \link{calculateEchoDepth} must be run first,
# and it requires a large amount of data not stored in the package
\dontrun{
study <- calculateEchoDepth(study, wav='path/to/wavFiles')
study <- runIciReview(x)
}

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
