% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFiles.R
\name{updateFiles}
\alias{updateFiles}
\title{Update Location of Binary and Databases Files}
\usage{
updateFiles(x, bin = NULL, db = NULL, quiet = FALSE)
}
\arguments{
\item{x}{an \linkS4class{AcousticStudy} or \linkS4class{AcousticEvent}
object}

\item{bin}{folder containing updated binary file locations. If
\code{NULL} (default), user will be prompted to select a folder}

\item{db}{folder containing updated database file locations. If
\code{NULL} (default), user will be prompted to select a folder}

\item{quiet}{logical flag to print messages about success of replacement}
}
\value{
the same \linkS4class{AcousticStudy} and
  \linkS4class{AcousticEvent} object as \code{x} with
  updated file locations
}
\description{
Updates the stored locations of binary and database files in
  the \code{files} slots of all \linkS4class{AcousticStudy} and
  \linkS4class{AcousticEvent} objects. Typically used after changing
  computers, or if original data was on an external hard drive. If any
  missing files are not able to be located, they will be kept in the files
  slot so that this function can be run again
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
