% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEventClips.R
\name{writeEventClips}
\alias{writeEventClips}
\title{Create Wav Clips of Events}
\usage{
writeEventClips(x, buffer = 0.1, outDir = ".")
}
\arguments{
\item{x}{\linkS4class{AcousticStudy} object containing events to make wav clips for}

\item{buffer}{amount around each event to also include in the clip, in seconds}

\item{outDir}{directory to write clips to, defaults to current directory}
}
\value{
A vector of file names for the wav clips that were successfully
  created, any that were not able to be written will be \code{NA}
}
\description{
Creates an audio clip containing sounds from an event
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
