% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEventClips.R
\name{writeEventClips}
\alias{writeEventClips}
\title{Create Wav Clips of Events}
\usage{
writeEventClips(x, buffer = 0.1, outDir = ".")
}
\arguments{
\item{x}{\linkS4class{AcousticStudy} object containing events to make wav clips for}

\item{buffer}{amount around each event to also include in the clip, in seconds}

\item{outDir}{directory to write clips to, defaults to current directory}
}
\value{
A vector of file names for the wav clips that were successfully
  created, any that were not able to be written will be \code{NA}. Note
  that currently this can only write clips with up to 2 channels.
}
\description{
Creates audio clips containing sounds from events
}
\examples{

data(exStudy)
recs <- system.file('extdata', 'Recordings', package='PAMpal')
exStudy <- addRecordings(exStudy, folder=recs, log=FALSE, progress=FALSE)
tdir <- tempdir()
wavs <- writeEventClips(exStudy, outDir=tdir)
file.remove(wavs)

}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
