% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEventClips.R
\name{writeEventClips}
\alias{writeEventClips}
\title{Create Wav Clips of Events}
\usage{
writeEventClips(
  event,
  wavFolder = NULL,
  buffer = 0.1,
  format = c("pamguard", "soundtrap"),
  log = NULL
)
}
\arguments{
\item{event}{\linkS4class{AcousticEvent} or \linkS4class{AcousticStudy}
object containing events to make wav clips for}

\item{wavFolder}{the folder where the wav files are located}

\item{buffer}{amount around each event to also include in the clip, in seconds}

\item{format}{either \code{'pamguard'} or \code{'soundtrap'} specifying
how the original audio files were created. The names of these files must
not have been changed, the original formatting is used to parse out the
start times of the audio files}

\item{log}{optional location of SoundTrap XML log files, only relevant for
\code{format='soundtrap'}}
}
\value{
A vector of file names for the wav clips that were successfully
  created, any that were not able to be written will be \code{NA}
}
\description{
Creates an audio clip containing sounds from an event
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
